--- scripts/modules/gun_system/__entry__.ts
@@ -1,4 +1,0 @@
-import './processors/combat/GunFireSystem'
-import './processors/combat/BulletSystem'
-import './processors/combat/GrenadeSystem'
-import './processors/combat/GunReloadSystem'
