+++ scripts/modules/weapon/__entry__.ts
@@ -1,0 +1,4 @@
+import './systems/combat/GunFireSystem'
+import './systems/combat/BulletSystem'
+import './systems/combat/GrenadeSystem'
+import './systems/combat/GunReloadSystem'
