+++ scripts/modules/weapon/actors/entity/Flashbang.ts
