+++ scripts/modules/weapon/actors/entity/SmokeGrenade.ts
