+++ scripts/modules/weapon/actors/item/Glock17.ts
