+++ scripts/modules/weapon/components/Component.ts
