+++ scripts/modules/weapon/components/EntityComponent.ts
