+++ scripts/modules/weapon/components/GrenadeComponent.ts
