+++ scripts/modules/weapon/components/ItemComponent.ts
