+++ scripts/modules/weapon/components/bullet/BulletComponent.ts
