+++ scripts/modules/weapon/components/bullet/BulletDamageComponent.ts
