+++ scripts/modules/weapon/components/gun/GunComponent.ts
