+++ scripts/modules/weapon/components/gun/GunFireComponent.ts
