+++ scripts/modules/weapon/components/gun/GunMagazineComponent.ts
