+++ scripts/modules/weapon/components/gun/GunOffsetComponent.ts
