+++ scripts/modules/weapon/systems/combat/BulletSystem.ts
@@ -1,4 +1,4 @@
-import { BulletActorTable } from "../../types/ActorTypeTables";
+import { BulletActorTable } from "../../actors/ActorTypeTables";
 import { spawnDummyEntity } from "../../../../utils/Utils";
 import { GunTypeEnum } from "../../types/Enums";
