+++ scripts/modules/weapon/types/Combat.d.ts
