+++ scripts/modules/weapon/types/Enums.ts
